+++ __init__.py
@@ -14,7 +14,7 @@ from pathlib import Path
 
 from .provider import PrebextorProvider
 
-__version__ = "1.0.1"
+__version__ = "1.0.2"
 __all__ = ["PrebextorProvider"]